pyproject.toml
[project]
name = "dev-sr-preperation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.24.0",
    "azure-storage-blob>=12.26.0",
    "jupyter>=1.1.1",
    "notebook>=7.4.5",
    "numpy>=2.3.2",
    "openai>=1.108.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
]
